Update CLI to not reformat the input file
Adjust the implementation of the CLI to leverage the `actions` output
list from the main library instead of the `workflow` output to rewrite
the input file. In particular, this constructs are regular expression
to find `uses:` directives to replace and replaces them (with a fixed
format). The result is that the input file isn't reformatted, apart from
the `uses:` directives.

The implementation roughly works as follows. For a given file, the main
library is invoked which returns an array of actions. For each action it
will construct a regular expression that matches `uses:` directives for
that action and then uses `"".replace` to replace the old action version
with the new one. In particular, it starts matching at `uses:`, then any
amount of whitespace (in a group so it can be preserved) then the action
as `<owner>/<repo>(/<path>)`, then the ref as `@<ref>` and finally any
trailing comment. This is replaced by `uses:`, then the whitespace using
`$1`, then the action again as `<owner>/<repo>(/<path>)`, then the new
ref as `@<new-ref>` followed by the (new) comment per the CLI config.

Currently this is implemented in for the CLI only. It might make sense
to lift this into the main library instead and use this rewrite strategy
for the `workflow` output instead of using `workflow.toString` (index.js
line 52).

Also note that this duplicates the logic for the default comment from
replaceActions.js line 93-96. This should probably be fixed prior to
releasing this.