Implement replaceActions in a text-based search-and-replace manner
Update the implementation of the replaceActions functionality from an
AST-based transformation to a text-based transformation using regular
expression to search and replace actions. This change was motivated by
the need to avoid changing the formatting of the YAML input.

The existing test suite for this part of the code base has not changed,
but did need to be updated to account for the new API of replaceActions.
On top of that, the test suite for replaceActions has been expanded with
several formatting related scenarios, mostly related to spacing around
the uses value as well as indentation.

This change eliminated the need for the `--yaml-line-width` and
`--yaml-null-str` inputs. They're removed from the code base internally,
but I kept them on the CLI to avoid breaking users that expect those
flags to exist - I just changed the help text to mark them as
deprecated.