Update CLI to not reformat input files (#158)
* Update CLI to not reformat the input file

Adjust the implementation of the CLI to leverage the `actions` output
list from the main library instead of the `workflow` output to rewrite
the input file. In particular, this constructs are regular expression
to find `uses:` directives to replace and replaces them (with a fixed
format). The result is that the input file isn't reformatted, apart from
the `uses:` directives.

The implementation roughly works as follows. For a given file, the main
library is invoked which returns an array of actions. For each action it
will construct a regular expression that matches `uses:` directives for
that action and then uses `"".replace` to replace the old action version
with the new one. In particular, it starts matching at `uses:`, then any
amount of whitespace (in a group so it can be preserved) then the action
as `<owner>/<repo>(/<path>)`, then the ref as `@<ref>` and finally any
trailing comment. This is replaced by `uses:`, then the whitespace using
`$1`, then the action again as `<owner>/<repo>(/<path>)`, then the new
ref as `@<new-ref>` followed by the (new) comment per the CLI config.

Currently this is implemented in for the CLI only. It might make sense
to lift this into the main library instead and use this rewrite strategy
for the `workflow` output instead of using `workflow.toString` (index.js
line 52).

Also note that this duplicates the logic for the default comment from
replaceActions.js line 93-96. This should probably be fixed prior to
releasing this.

* Improve expression to update workflow files

Update the regular expression:
1. To allow for spacing between `uses` and `:` to support workflows
   formatted as such.
2. To require at least one space after `uses:`, which is required by the
   YAML syntax. Requiring this in the expression avoids accidental
   matching elsewhere in the file.
3. Preserve spacing between the value and the pin comment. To achieve
   this I had to split the expression for updating a pinned value and
   adding a new pin comment.

I also checked whether this will update the workflow if there's no final
newline after a `uses:` directive and it already does so that did not
require an update.

* Implement replaceActions in a text-based search-and-replace manner

Update the implementation of the replaceActions functionality from an
AST-based transformation to a text-based transformation using regular
expression to search and replace actions. This change was motivated by
the need to avoid changing the formatting of the YAML input.

The existing test suite for this part of the code base has not changed,
but did need to be updated to account for the new API of replaceActions.
On top of that, the test suite for replaceActions has been expanded with
several formatting related scenarios, mostly related to spacing around
the uses value as well as indentation.

This change eliminated the need for the `--yaml-line-width` and
`--yaml-null-str` inputs. They're removed from the code base internally,
but I kept them on the CLI to avoid breaking users that expect those
flags to exist - I just changed the help text to mark them as
deprecated.

* Handle RegExp meta-characters in action IDs and action versions

* Remove --yaml* options as formatting is now preserved

---------

Co-authored-by: Michael Heap <m@michaelheap.com>